requirements.txt
astropy==4.0.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
